feat: tests for enrich
