+++ test/event.test.js
@@ -1,9 +1,7 @@
 require('./test_helper'); 
 
-const HASH_RESULT = 'NEW HASH FOR THE EVENT';
 const EVENT_UUID_RESULT = 'NEW UUID FOR THE EVENT';
 
-const omit = require('lodash.omit');
 const AWS = require('aws-sdk');
 const kinesis = new AWS.Kinesis({ region: 'eu-west-1' });
 
@@ -121,74 +119,5 @@ describe('#emitEvent', () => {
           });
           clock.restore();
         });
-  });
-
-  describe('when creating a PartitionKey', () => {
-    describe('when no identifier is passed', () => {
-      const event = { key: 'value' };
-      
-      // it('emits event with random PartitionKey', () => {
-      //   toStringStub.withArgs('hex').returns(null);
-
-      //   emitEvent(kinesis, event, config);
-      //   expect(putRecordStub).to.have.been.calledWithMatch(
-      //     { PartitionKey: HASH_RESULT });
-      //   });
-    });
-
-    const eventWithAllIdentifiers = {
-      key: 'value',
-      uuid: 'user_uuid',
-      tracking_uuid: 'user_tracking_uuid',
-      meta: {
-        udid: 'mobile_device_udid',
-        event_uuid: 'event_uuid'
-      }
-    };
-
-    context('when uuid, tracking_uuid, meta.udid and meta.event_uuid are set', () => {
-      it('emits event with PartitionKey == event.uuid', () => {
-        const event = eventWithAllIdentifiers;
-        emitEvent(kinesis, event, config);
-
-        expect(putRecordStub).to.have.been.calledWithMatch(
-          { PartitionKey: 'user_uuid' });
-      });
-    });
-
-    context('when tracking_uuid, meta.udid and meta.event_uuid are set', () => {
-      const event = omit(eventWithAllIdentifiers, ['uuid']);
-
-      it('emits event with PartitionKey == event.tracking_uuid', () => {
-        emitEvent(kinesis, event, config);
-        expect(putRecordStub).to.have.been.calledWithMatch(
-          { PartitionKey: 'user_tracking_uuid' });
-      });
-    });
-
-    context('when meta.udid and meta.event_uuid are set', () => {
-      const event = omit(eventWithAllIdentifiers, ['uuid', 'tracking_uuid']);
-
-      it('emits event with PartitionKey == event.meta.udid', () => {
-        emitEvent(kinesis, event, config);
-        expect(putRecordStub).to.have.been.calledWithMatch(
-          { PartitionKey: 'mobile_device_udid' });
-      });
-    });
-
-    context('when meta.event_uuid is set', () => {
-      const event = {
-        key: 'value',
-        meta: {
-          event_uuid: 'event_uuid'
-        }
-      };
-
-      it('emits event with PartitionKey == event.meta.event_uuid', () => {
-        emitEvent(kinesis, event, config);
-        expect(putRecordStub).to.have.been.calledWithMatch(
-          { PartitionKey: 'event_uuid' });
-      });
-    });
   });
 });
